+++ weave-js/package.json
@@ -34,7 +34,6 @@
     "@emotion/styled": "^11.11.0",
     "@floating-ui/react": "^0.24.5",
     "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
     "@monaco-editor/react": "^4.3.1",
     "@mui/icons-material": "^5.14.12",
     "@mui/material": "^6.0.2",
@@ -95,7 +94,6 @@
     "react": "^17.0.2",
     "react-app-polyfill": "^3.0.0",
     "react-base-table": "^1.12.0",
-    "react-beautiful-dnd": "^13.1.1",
     "react-cytoscapejs": "^1.2.1",
     "react-datetime": "^3",
     "react-diff-viewer": "^3.1.1",
